# docs/openapi.full.yaml
openapi: 3.0.3
info:
  title: Biometric Face Attendance API (Full)
  version: 1.1.0
  description: |
    Exhaustive endpoint specification for the backend route surface.
    Includes all discovered routes from auth, admin, lecturer, student, and recognition blueprints.
  x-last-reviewed: 2026-04-14
servers:
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    jwtCookie:
      type: apiKey
      in: cookie
      name: access_token_cookie
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    GenericObject:
      type: object
      additionalProperties: true
...
    AdminPaper:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        course_id: { type: string }
        lecturer_id:
          type: string
          nullable: true
        semester:
          type: integer
          nullable: true
        total_classes: { type: integer }
      additionalProperties: true
...
    PaginatedPapersResponse:
      type: object
      required: [items, total, page, per_page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminPaper' }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
    AdminPapersListResponse:
      oneOf:
        - type: array
          items: { $ref: '#/components/schemas/AdminPaper' }
        - $ref: '#/components/schemas/PaginatedPapersResponse'
...
security:
  - jwtCookie: []
paths:
...
  /api/admin/papers:
    get:
      tags: [AdminPapers]
      operationId: adminListPapers
      responses:
        '200': { description: Papers, content: { application/json: { schema: { $ref: '#/components/schemas/AdminPapersListResponse' } } } }
    post:
      tags: [AdminPapers]
      operationId: adminCreatePaper
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code, course_id, lecturer_id, semester]
              properties:
                name: { type: string }
                code: { type: string }
                course_id: { type: string }
                lecturer_id: { type: string }
                semester: { type: integer }
      responses:
        '201': { description: Paper created, content: { application/json: { schema: { $ref: '#/components/schemas/AdminPaper' } } } }
  /api/admin/papers/{pid}:
    put:
      tags: [AdminPapers]
      operationId: adminUpdatePaper
      security:
        - jwtCookie: []
          csrfHeader: []
      parameters:
        - in: path
          name: pid
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericObject' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
    delete:
      tags: [AdminPapers]
      operationId: adminDeletePaper
      security:
        - jwtCookie: []
          csrfHeader: []
      parameters:
        - in: path
          name: pid
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted/deactivated, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
  /api/admin/papers/bulk-assign:
    post:
      tags: [AdminPapers]
      operationId: adminBulkAssignPaper
      security:
        - jwtCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paper_id: { type: string }
                paper_ids: { type: array, items: { type: string } }
                lecturer_id: { type: string }
                course_id: { type: string }
                student_ids: { type: array, items: { type: string } }
      responses:
        '200': { description: Assignment completed, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
...